scrapers/carnationcanada/requirements.txt
boto3==1.20.28
botocore==1.23.28
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
jmespath==0.10.0
pandas==1.3.5
psycopg2==2.9.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
undetected-chromedriver==3.1.3
python-decouple==3.5

